script builds, stops, runs

+++ make.sh
@@ -5,6 +5,8 @@ SRC="src"
 API="deploy/api"
 CELERY="deploy/celery"
 
+
+
 git pull
 
 if [ "api" == $TARGET ]; then
@@ -21,6 +23,9 @@ else
 	exit 0
 fi
 
+CID=$(sudo docker ps | grep "$DOCKERBUILD" | awk '{print $1}')
+echo $CID
+
 if [ -d "$SRC" ]; then
 	echo ">> Cleaning up"
 	rm -rf $SRC
@@ -30,10 +35,35 @@ echo ">> Adding source"
 cp -r ../../src .
 
 echo ">> Running Docker Build"
-docker build -t $DOCKERBUILD .
+# docker build -t $DOCKERBUILD .
+
+sudo docker build -t $DOCKERBUILD . | tee /tmp/docker_build_result.log
+RESULT=$(cat /tmp/docker_build_result.log | tail -n 1)
+if [[ "$RESULT" != *Successfully* ]];
+then
+  exit -1
+fi
 
 
 if [ -d "$SRC" ]; then
 	echo ">> Cleaning up"
 	rm -rf $SRC
+fi
+
+
+echo '>> Stopping old container'
+if [ "$CID" != "" ];
+then
+  sudo docker stop $CID
+fi
+
+if [ "api" == $TARGET ]; then
+
+	docker run -d -p 5001:5000 --link mongodb:mongodb --link redis:redis \
+    	-v /var/log/containers/dotmarks-api:/var/log ipedrazas/dotmarks-api
+
+elif [ "celery" == $TARGET ]; then
+	docker run -d --link redis:redis --link mongodb:mongodb \
+		-v /var/log/containers/dotmarks-celery:/log ipedrazas/dotmarks-celery
+
 fi
